requirements.txt
# pip freeze  - in termimal -> shows all dependencies we have installed
flask
python-dotenv
pymongo[srv]
gunicorn


# pip install -r requirements.txt

# flask run 
# OR
# gunicorn "app:create_app()"
